split read and write
